README file
Name of project 
JEE Image Search.
Names of team members 
 Jessica Jimenez,Evert Rodriguez, Eric Tinoco
Class: 
CST205_01: Multimedia Design and Programming.
Date: 
December 10, 2017
How to run program:
The Program will run through a GUI. When The GUI appears the user will be asked “What Image would you like to search for?” with audio. The user will speak to the GUI to search for the image. The GUI will get the users speech input and transform it into a text. The text will be used to search for the image using flicker. 
Link to GitHub repository:
https://github.com/tinoco86/TeamJEE
Future work:
In future work we would like to give the user more images. We would also like the user to be able to change the image. For example, the user would be able to change the image filter or size.